+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransactionType.java
@@ -3,7 +3,7 @@ package com.concordium.sdk.transactions;
 import lombok.Getter;
 
 // Types must match https://github.com/Concordium/concordium-base/blob/main/haskell-src/Concordium/Types/Execution.hs
-enum TransactionType {
+public enum TransactionType {
     DEPLOY_MODULE((byte) 0),
     INITIALIZE_SMART_CONTRACT_INSTANCE((byte) 1),
     UPDATE_SMART_CONTRACT_INSTANCE((byte) 2),